adding SnapshotDensityMatrix and tests
typo in test

test

test fixes

uncommenting important things

test

test

test

tidying up

renaming files; testing density_matrix snapshot

renamed file

renamed file

small typo fix

fixed import in __init__.py
