+++ qiskit/providers/aer/__init__.py
@@ -27,7 +27,7 @@ from .aerjob import AerJob
 from .aererror import AerError
 from .backends import *
 from . import noise
-from . import snapshots
+from . import aer_snapshots
 from . import utils
 from .version import __version__
